refactor(meeting-bot): split configured host wrapper
